requirements.txt
faiss-cpu==1.8.0
langchain-community==0.0.34
langchain==0.1.16
openai==1.23.6
langchain-openai==0.1.4
streamlit==1.33.0
pdf2image==1.17.0
python-dotenv==1.0.1